refactor the code a bit and also improve the fade out fade in effect and also added icons and manifest.json
