+++ components/CreateAnimation/Hover/index.tsx
@@ -4,11 +4,11 @@ import gsap from 'gsap';
 
 const Hover: React.FC = ({ children }) => {
 	const onEnter = (event: { currentTarget: gsap.TweenTarget }) => {
-		gsap.to(event.currentTarget, { y: -20, scale: 1.2 });
+		gsap.to(event.currentTarget, { y: -10, scale: 1.2 });
 	};
 
 	const onLeave = (event: { currentTarget: gsap.TweenTarget }) => {
-		gsap.to(event.currentTarget, { y: 0, scale: 0.8 });
+		gsap.to(event.currentTarget, { y: 0, scale: 1 });
 	};
 
 	return (
