+++ components/Layout/Navigation/index.tsx
@@ -15,8 +15,7 @@ const Navigation = () => {
 		home = '/',
 		about = '/about',
 		portfolio = '/portfolio',
-	};
-	
+	}
 
 	return (
 		<S.NavigationWrapper>
