+++ pages/portfolio.tsx
@@ -13,11 +13,10 @@ import PageLayout from '../components/Layout';
 const PortfolioPage: NextPage = () => {
 	const { t } = useTranslation('portfolio');
 
-
 	return (
 		<PageLayout title={t('title')}>
 			<S.PageWrapper>
-				<div >
+				<div>
 					<h2>portfolioing vietnam</h2>
 				</div>
 			</S.PageWrapper>
