+++ styles/pages/portfolio.ts
@@ -1,9 +1,8 @@
-import { device } from 'helpers/mediaQueries';
+import { device } from 'utils/mediaQueries';
 import styled from 'styled-components';
 
-export const PageWrapper = styled.main `
-
-${device.tablet} {
-	width: 10rem;
+export const PageWrapper = styled.main`
+	${device.tablet} {
+		width: 10rem;
 	}
-`
+`;
